Fix bug 87279 / LP TODO (Error log pattern searches broken in main.mysql_upgrade_test)
The testcase greps the default server error log for expected
patterns. The default log contains all the error log from the MTR run
on the same worker so far. This is bad because now the searches can
find output of the preceding tests. This breaks --repeat option, retry
on failure, and should any previous test caused same output in the
server log, the corresponding part of the test stops working and just
passes silently.

Fix by restarting the server with the custom error log, grepping that
instead, resetting to the default error log after the last grep, and
deleting the custom log.