Fix bug 87282 / LP TODO (WL#9499 can cause extra warnings on every crashing testcase)
Each crashing testcase with a low probability may fail its warning
check with an extra warning about the tablespace map file being zero
length. This could be caused by the InnoDB background thread making
checkpoints in the background. Each checkpoint writes out a tablespace
map file after WL#9499, and before writing out, a file is truncated to
zero. If a crash in another thread happens at this point, the crash
recovery will produce this extra warning.

Fix by suppressing the warning. To avoid suppressing it globally, add
the suppression to include/not_crashrep.inc, which should be included
by the crashing testcases. Include this file in some of the crashing
testcases that do not include it already.