PS-6789 (Merge MySQL 8.0.20): Fixed rpl.percona_rpl_encryption_master_binlog_ps_encrypted MTR test case
https://jira.percona.com/browse/PS-6789

Fixed upstream problem in the 'include/show_events.inc' MTR include file with
invalid calculation of the second's event position in case when custom
'$binlog_file' is specified.

The problem revealed itself when 8.0.19 prerecorded binlog (with shorter
'Format_description_event') was used on a 8.0.20 server (where
'Format_description_event' is 1 byte longer because of the new
'Transaction_payload_event').