Added --after-failure-hook MTR command line option
Added --after-failure-hook=COMMAND MTR command line option that
allows to specify custom command that will be executed upon test
failure.

Currently it is used by MTR runs confiured to use RocksDB in
combination with ZenFS on zoned block devices (via
'--rocksdb-fs-uri=zenfs://...' option).
The purpose of this option is to perform ZenFS cleanup when MTR
encounters a failing test - in this case, just cleaning mysqld
data directory is not enough as data files may still remain on
ZenFS.

For instance,
./mtr --suite=rocksdb --mysql=--rocksdb-fs-uri=zenfs://dev:nullb0 \
  --after-failure-hook="rm -rf /var/lib/mysql_zenfs_aux; zenfs mkfs --zbd=nullb0 --aux_path=/var/lib/mysql_zenfs_aux --force"