Fixed memory corruption problem in DIRECT_CHECK logic on ZenFS
Currently on ZenFS it is not safe to call 'DeleteFile()' for a file
for which 'WritableFile' object was created and is atill alive.

This used to have memory corruption problems when such 'WritableFile'
object was getting out of scope and its destructor was called causing
freeing resources already freed by the 'DeleteFile()'.

As a workaround, 'DIRECT_CHECK' logic reworked so that 'DeleteFile()'
is now called after 'WritableFile' is destroyed.