PS-8728 feature: Align Percona's Data Masking functionality with Oracle 8.0.33 (patch 8)
https://jira.percona.com/browse/PS-8728

The following functions when invoked with NULL as a primary (the first)
argument instead of reporting an error now immediately return NULL:
* mask_canada_sin()
* mask_iban()
* mask_inner()
* mask_outer()
* mask_pan()
* mask_pan_relaxed()
* mask_ssn()
* mask_uk_nin()
* mask_uuid()
* gen_blocklist()

This helps a lot in scenarios when a masking function is used in SELECT
statements on fields that may include 'NULL' values.
For instance,
SELECT name, mask_pan_relaxed(cc_number) FROM tbl;
where 'cc_number' field in the 'tbl' table is not declared as
'NOT NULL'.

Squash-with: d2e2271