Refactored string constants
Removed static keyword because the declarations are made in the anonymous namespace.

Everything made std::string_view except for those that need to be used in c-style interfaces
(need to be NUL-terminated).

It does not make sense to throw 'udf_exception' with custom error code ( say, ER_EXCESS_ARGUMENTS}
from the constructors.
Because of the the Server UDF lifecycle, error reported from the 'xxx_init()' function
will be always be reported as ER_CANT_INITIALIZE_UDF and there is not way to customize this.
However, 'mysqlpp::udf_exception' can be thrown from the 'calculate()' methods.
In this case, the error code can be customized.

Currently, changing all throws to 'raise<std::invalid_argument>(err_msg)'.

Introduced 'rise<>()' template helper function for throwing exceptions.

Extended the scope of the anonymous namespace to include all 'xxx_impl' class definitions.

Fixed "extra ';' outside of a function is incompatible with C++98 [-Wc++98-compat-extra-semi]"
warnings.

Fixed some misprints.